breaks out nerf utils
