+++ examples/nerf/train_nerf.py
@@ -16,7 +16,8 @@ import wandb
 from fastcore.script import *
 from wandb.keras import WandbModelCheckpoint, WandbMetricsLogger
 
-# from configs.utils import import_config
+from nerf_utils import *
+
 @call_parse
 def main(
         wandb_run_id:Param("wandb run id from", str)=None,
@@ -39,7 +40,8 @@ def main(
     EPOCHS = wandb.config.EPOCHS
     POS_ENCODE_DIMS = wandb.config.POS_ENCODE_DIMS
     SAVE_FREQ = wandb.config.SAVE_FREQ
-    
+    MODEL_NAME = wandb.config.MODEL_NAME
+
 
     # Download the data if it does not already exist.
     file_name = "tiny_nerf_data.npz"
@@ -52,116 +54,6 @@ def main(
     im_shape = images.shape
     (num_images, H, W, _) = images.shape
     (poses, focal) = (data["poses"], data["focal"])
-
-    # # Plot a random image from the dataset for visualization.
-    # plt.imshow(images[np.random.randint(low=0, high=num_images)])
-    # plt.show()
-
-
-    def encode_position(x):
-        """Encodes the position into its corresponding Fourier feature.
-        Args:
-            x: The input coordinate.
-        Returns:
-            Fourier features tensors of the position.
-        """
-        positions = [x]
-        for i in range(POS_ENCODE_DIMS):
-            for fn in [tf.sin, tf.cos]:
-                positions.append(fn(2.0**i * x))
-        return tf.concat(positions, axis=-1)
-
-
-    def get_rays(height, width, focal, pose):
-        """Computes origin point and direction vector of rays.
-        Args:
-            height: Height of the image.
-            width: Width of the image.
-            focal: The focal length between the images and the camera.
-            pose: The pose matrix of the camera.
-        Returns:
-            Tuple of origin point and direction vector for rays.
-        """
-        # Build a meshgrid for the rays.
-        i, j = tf.meshgrid(
-            tf.range(width, dtype=tf.float32),
-            tf.range(height, dtype=tf.float32),
-            indexing="xy",
-        )
-
-        # Normalize the x axis coordinates.
-        transformed_i = (i - width * 0.5) / focal
-
-        # Normalize the y axis coordinates.
-        transformed_j = (j - height * 0.5) / focal
-
-        # Create the direction unit vectors.
-        directions = tf.stack([transformed_i, -transformed_j, -tf.ones_like(i)], axis=-1)
-
-        # Get the camera matrix.
-        camera_matrix = pose[:3, :3]
-        height_width_focal = pose[:3, -1]
-
-        # Get origins and directions for the rays.
-        transformed_dirs = directions[..., None, :]
-        camera_dirs = transformed_dirs * camera_matrix
-        ray_directions = tf.reduce_sum(camera_dirs, axis=-1)
-        ray_origins = tf.broadcast_to(height_width_focal, tf.shape(ray_directions))
-
-        # Return the origins and directions.
-        return (ray_origins, ray_directions)
-
-
-    def render_flat_rays(ray_origins, ray_directions, near, far, num_samples, rand=False):
-        """Renders the rays and flattens it.
-        Args:
-            ray_origins: The origin points for rays.
-            ray_directions: The direction unit vectors for the rays.
-            near: The near bound of the volumetric scene.
-            far: The far bound of the volumetric scene.
-            num_samples: Number of sample points in a ray.
-            rand: Choice for randomising the sampling strategy.
-        Returns:
-        Tuple of flattened rays and sample points on each rays.
-        """
-        # Compute 3D query points.
-        # Equation: r(t) = o+td -> Building the "t" here.
-        t_vals = tf.linspace(near, far, num_samples)
-        if rand:
-            # Inject uniform noise into sample space to make the sampling
-            # continuous.
-            shape = list(ray_origins.shape[:-1]) + [num_samples]
-            noise = tf.random.uniform(shape=shape) * (far - near) / num_samples
-            t_vals = t_vals + noise
-
-        # Equation: r(t) = o + td -> Building the "r" here.
-        rays = ray_origins[..., None, :] + (
-            ray_directions[..., None, :] * t_vals[..., None]
-        )
-        rays_flat = tf.reshape(rays, [-1, 3])
-        rays_flat = encode_position(rays_flat)
-        return (rays_flat, t_vals)
-
-
-    def map_fn(pose):
-        """Maps individual pose to flattened rays and sample points.
-        Args:
-            pose: The pose matrix of the camera.
-        Returns:
-            Tuple of flattened rays and sample points corresponding to the
-            camera pose.
-        """
-        (ray_origins, ray_directions) = get_rays(height=H, width=W, focal=focal, pose=pose)
-        (rays_flat, t_vals) = render_flat_rays(
-            ray_origins=ray_origins,
-            ray_directions=ray_directions,
-            near=2.0,
-            far=6.0,
-            num_samples=NUM_SAMPLES,
-            rand=True,
-        )
-        return (rays_flat, t_vals)
-
 
     # Create the training split.
     split_index = int(num_images * 0.8)
@@ -196,191 +88,10 @@ def main(
         .prefetch(AUTO)
     )
 
-
-    def get_nerf_model(num_layers, num_pos):
-        """Generates the NeRF neural network.
-        Args:
-            num_layers: The number of MLP layers.
-            num_pos: The number of dimensions of positional encoding.
-        Returns:
-            The `tf.keras` model.
-        """
-        inputs = keras.Input(shape=(num_pos, 2 * 3 * POS_ENCODE_DIMS + 3))
-        x = inputs
-        for i in range(num_layers):
-            x = layers.Dense(units=64, activation="relu")(x)
-            if i % 4 == 0 and i > 0:
-                # Inject residual connection.
-                x = layers.concatenate([x, inputs], axis=-1)
-        outputs = layers.Dense(units=4)(x)
-        return keras.Model(inputs=inputs, outputs=outputs)
-
-
-    def render_rgb_depth(model, rays_flat, t_vals, rand=True, train=True):
-        """Generates the RGB image and depth map from model prediction.
-        Args:
-            model: The MLP model that is trained to predict the rgb and
-                volume density of the volumetric scene.
-            rays_flat: The flattened rays that serve as the input to
-                the NeRF model.
-            t_vals: The sample points for the rays.
-            rand: Choice to randomise the sampling strategy.
-            train: Whether the model is in the training or testing phase.
-        Returns:
-            Tuple of rgb image and depth map.
-        """
-        # Get the predictions from the nerf model and reshape it.
-        if train:
-            predictions = model(rays_flat)
-        else:
-            predictions = model.predict(rays_flat)
-        predictions = tf.reshape(predictions, shape=(BATCH_SIZE, H, W, NUM_SAMPLES, 4))
-
-        # Slice the predictions into rgb and sigma.
-        rgb = tf.sigmoid(predictions[..., :-1])
-        sigma_a = tf.nn.relu(predictions[..., -1])
-
-        # Get the distance of adjacent intervals.
-        delta = t_vals[..., 1:] - t_vals[..., :-1]
-        # delta shape = (num_samples)
-        if rand:
-            delta = tf.concat(
-                [delta, tf.broadcast_to([1e10], shape=(BATCH_SIZE, H, W, 1))], axis=-1
-            )
-            alpha = 1.0 - tf.exp(-sigma_a * delta)
-        else:
-            delta = tf.concat(
-                [delta, tf.broadcast_to([1e10], shape=(BATCH_SIZE, 1))], axis=-1
-            )
-            alpha = 1.0 - tf.exp(-sigma_a * delta[:, None, None, :])
-
-        # Get transmittance.
-        exp_term = 1.0 - alpha
-        epsilon = 1e-10
-        transmittance = tf.math.cumprod(exp_term + epsilon, axis=-1, exclusive=True)
-        weights = alpha * transmittance
-        rgb = tf.reduce_sum(weights[..., None] * rgb, axis=-2)
-
-        if rand:
-            depth_map = tf.reduce_sum(weights * t_vals, axis=-1)
-        else:
-            depth_map = tf.reduce_sum(weights * t_vals[:, None, None], axis=-1)
-        return (rgb, depth_map)
-
-
-    class NeRF(keras.Model):
-        def __init__(self, nerf_model):
-            super().__init__()
-            self.nerf_model = nerf_model
-
-        def compile(self, optimizer, loss_fn):
-            super().compile()
-            self.optimizer = optimizer
-            self.loss_fn = loss_fn
-            self.loss_tracker = keras.metrics.Mean(name="loss")
-            self.psnr_metric = keras.metrics.Mean(name="psnr")
-
-        def train_step(self, inputs):
-            # Get the images and the rays.
-            (images, rays) = inputs
-            (rays_flat, t_vals) = rays
-
-            with tf.GradientTape() as tape:
-                # Get the predictions from the model.
-                rgb, _ = render_rgb_depth(
-                    model=self.nerf_model, rays_flat=rays_flat, t_vals=t_vals, rand=True
-                )
-                loss = self.loss_fn(images, rgb)
-
-            # Get the trainable variables.
-            trainable_variables = self.nerf_model.trainable_variables
-
-            # Get the gradeints of the trainiable variables with respect to the loss.
-            gradients = tape.gradient(loss, trainable_variables)
-
-            # Apply the grads and optimize the model.
-            self.optimizer.apply_gradients(zip(gradients, trainable_variables))
-
-            # Get the PSNR of the reconstructed images and the source images.
-            psnr = tf.image.psnr(images, rgb, max_val=1.0)
-
-            # Compute our own metrics
-            self.loss_tracker.update_state(loss)
-            self.psnr_metric.update_state(psnr)
-            return {"loss": self.loss_tracker.result(), "psnr": self.psnr_metric.result()}
-
-        def test_step(self, inputs):
-            # Get the images and the rays.
-            (images, rays) = inputs
-            (rays_flat, t_vals) = rays
-
-            # Get the predictions from the model.
-            rgb, _ = render_rgb_depth(
-                model=self.nerf_model, rays_flat=rays_flat, t_vals=t_vals, rand=True
-            )
-            loss = self.loss_fn(images, rgb)
-
-            # Get the PSNR of the reconstructed images and the source images.
-            psnr = tf.image.psnr(images, rgb, max_val=1.0)
-
-            # Compute our own metrics
-            self.loss_tracker.update_state(loss)
-            self.psnr_metric.update_state(psnr)
-            return {"loss": self.loss_tracker.result(), "psnr": self.psnr_metric.result()}
-
-        @property
-        def metrics(self):
-            return [self.loss_tracker, self.psnr_metric]
-
-        # ADDED
-        def save_weights(self, filepath, *args, **kwargs):
-            """While saving the weights, we simply save the weights of the DCE-Net"""
-            self.nerf_model.save_weights(filepath, *args, **kwargs)
-
-        def load_weights(self, filepath, *args, **kwargs):
-            """While loading the weights, we simply load the weights of the DCE-Net"""
-            self.nerf_model.load_weights(filepath=filepath, *args, **kwargs)
-        
-        def save(self, filepath, *args, **kwargs):
-            self.nerf_model.save(filepath, *args, **kwargs)
-
-        def save_weights(self, filepath, *args, **kwargs):
-            self.nerf_model.save_weights(filepath, *args, **kwargs)
-        # END ADDED
-
     test_imgs, test_rays = next(iter(train_ds))
     test_rays_flat, test_t_vals = test_rays
 
     loss_list = []
-
-
-    class TrainMonitor(keras.callbacks.Callback):
-        def on_epoch_end(self, epoch, logs=None):
-            loss = logs["loss"]
-            loss_list.append(loss)
-            test_recons_images, depth_maps = render_rgb_depth(
-                model=self.model.nerf_model,
-                rays_flat=test_rays_flat,
-                t_vals=test_t_vals,
-                rand=True,
-                train=False,
-            )
-
-            # Plot the rgb, depth and the loss plot.
-            fig, ax = plt.subplots(nrows=1, ncols=3, figsize=(20, 5))
-            ax[0].imshow(keras.preprocessing.image.array_to_img(test_recons_images[0]))
-            ax[0].set_title(f"Predicted Image: {epoch:03d}")
-
-            ax[1].imshow(keras.preprocessing.image.array_to_img(depth_maps[0, ..., None]))
-            ax[1].set_title(f"Depth Map: {epoch:03d}")
-
-            ax[2].plot(loss_list)
-            ax[2].set_xticks(np.arange(0, EPOCHS + 1, 5.0))
-            ax[2].set_title(f"Loss Plot: {epoch:03d}")
-
-            fig.savefig(f"images/{epoch:03d}.png")
-            plt.show()
-            plt.close()
 
 
     num_pos = H * W * NUM_SAMPLES
@@ -402,20 +113,9 @@ def main(
         epochs=EPOCHS,
         callbacks=[TrainMonitor(), 
                    WandbMetricsLogger(log_freq="batch"),
-                   WandbModelCheckpoint(filepath='tmp', save_freq=SAVE_FREQ)],
+                   WandbModelCheckpoint(filepath=MODEL_NAME, save_freq=SAVE_FREQ)],
         steps_per_epoch=split_index // BATCH_SIZE,
     )
-
-
-    def create_gif(path_to_images, name_gif):
-        filenames = glob.glob(path_to_images)
-        filenames = sorted(filenames)
-        images = []
-        for filename in tqdm(filenames):
-            images.append(imageio.imread(filename))
-        kargs = {"duration": 0.25}
-        imageio.mimsave(name_gif, images, "GIF", **kargs)
-
 
     create_gif("images/*.png", "training.gif")
 
@@ -445,52 +145,6 @@ def main(
             keras.preprocessing.image.array_to_img(depth_map[..., None]), cmap="inferno"
         )
         ax[2].set_title("Depth Map")
-
-
-    def get_translation_t(t):
-        """Get the translation matrix for movement in t."""
-        matrix = [
-            [1, 0, 0, 0],
-            [0, 1, 0, 0],
-            [0, 0, 1, t],
-            [0, 0, 0, 1],
-        ]
-        return tf.convert_to_tensor(matrix, dtype=tf.float32)
-
-
-    def get_rotation_phi(phi):
-        """Get the rotation matrix for movement in phi."""
-        matrix = [
-            [1, 0, 0, 0],
-            [0, tf.cos(phi), -tf.sin(phi), 0],
-            [0, tf.sin(phi), tf.cos(phi), 0],
-            [0, 0, 0, 1],
-        ]
-        return tf.convert_to_tensor(matrix, dtype=tf.float32)
-
-
-    def get_rotation_theta(theta):
-        """Get the rotation matrix for movement in theta."""
-        matrix = [
-            [tf.cos(theta), 0, -tf.sin(theta), 0],
-            [0, 1, 0, 0],
-            [tf.sin(theta), 0, tf.cos(theta), 0],
-            [0, 0, 0, 1],
-        ]
-        return tf.convert_to_tensor(matrix, dtype=tf.float32)
-
-
-    def pose_spherical(theta, phi, t):
-        """
-        Get the camera to world matrix for the corresponding theta, phi
-        and t.
-        """
-        c2w = get_translation_t(t)
-        c2w = get_rotation_phi(phi / 180.0 * np.pi) @ c2w
-        c2w = get_rotation_theta(theta / 180.0 * np.pi) @ c2w
-        c2w = np.array([[-1, 0, 0, 0], [0, 0, 1, 0], [0, 1, 0, 0], [0, 0, 0, 1]]) @ c2w
-        return c2w
-
 
     rgb_frames = []
     batch_flat = []
